+++ functions/Get-TibberPriceInfo.ps1
@@ -8,26 +8,41 @@
         $response = Get-TibberPriceInfo -Last 10
         $maxPrice = $response | Sort-Object -Property total -Descending | Select-Object -First 1
         Write-Host "Max energy price, $($maxPrice.total) $($maxPrice.currency), starting at $(([DateTime]$maxPrice.startsAt).ToString('yyyy-MM-dd HH:mm')) [$($maxPrice.level)]"
+    .Example
+        $response = Get-TibberPriceInfo -IncludeToday
+        Write-Host "Today's energy prices: $($response | Out-String)"
+    .Example
+        $response = Get-TibberPriceInfo -IncludeTomorrow
+        Write-Host "Tomorrow's energy prices: $($response | Out-String)"
     .Link
         Invoke-TibberQuery
     .Link
         https://developer.tibber.com/docs/reference#priceinfo
     #>
-    [CmdletBinding(DefaultParameterSetName = '__None')]
+    [CmdletBinding(DefaultParameterSetName = '__AllParameterSets')]
     param (
         # Specifies the home Id, e.g. '96a14971-525a-4420-aae9-e5aedaa129ff'.
         [Parameter(Mandatory = $true, ParameterSetName = 'HomeId', ValueFromPipelineByPropertyName)]
         [Alias('Id')]
         [string] $HomeId,
 
-        # Specifies the resoluton of the results.
-        [Parameter(ValueFromPipelineByPropertyName)]
-        [ValidateSet('HOURLY', 'DAILY', 'WEEKLY', 'MONTHLY', 'ANNUAL')]
-        [string] $Resolution = 'HOURLY',
+        # Switch to exclude current energy price from the results.
+        [switch] $ExcludeCurrent,
+
+        # Switch to include today's energy price in the results.
+        [switch] $IncludeToday,
+
+        # Switch to include tomorrow's energy price in the results, available after 13:00 CET/CEST.
+        [switch] $IncludeTomorrow,
 
         # Specifies the number of nodes to include in results, counting back from the latest entry.
         [Parameter(ValueFromPipelineByPropertyName)]
         [int] $Last = 0,
+
+        # Specifies the resoluton of the results (applicable only when '-Last' is provided).
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [ValidateSet('HOURLY', 'DAILY', 'WEEKLY', 'MONTHLY', 'ANNUAL')]
+        [string] $Resolution = 'HOURLY',
 
         # Specifies the fields to return.
         # https://developer.tibber.com/docs/reference#price
@@ -48,9 +63,6 @@
     }
 
     process {
-        # Construct the GraphQL query arguments
-        $arguments = "resolution:$Resolution, last:$Last"
-
         # Construct the GraphQL query
         $query = "{ viewer{ "
         if ($PSCmdlet.ParameterSetName -eq 'HomeId') {
@@ -62,8 +74,23 @@
             $query += "$homeNode{ "
         }
         $query += "currentSubscription{ priceInfo{ "
-        $query += "current{ $($Fields -join ','), __typename }"
-        $query += "range($arguments){ nodes{ $($Fields -join ','),__typename } }"
+        $queryHasNoNode = $true
+        if (-Not $ExcludeCurrent.IsPresent) {
+            $query += "current{ $($Fields -join ','), __typename } "
+            $queryHasNoNode = $false
+        }
+        if ($IncludeToday.IsPresent) {
+            $query += "today{ $($Fields -join ','),__typename } "
+            $queryHasNoNode = $false
+        }
+        if ($IncludeTomorrow.IsPresent) {
+            $query += "tomorrow{ $($Fields -join ','),__typename } "
+            $queryHasNoNode = $false
+        }
+        if ($Last -gt 0 -Or $queryHasNoNode) {
+            $arguments = "resolution:$Resolution, last:$Last"
+            $query += "range($arguments){ nodes{ $($Fields -join ','),__typename } } "
+        }
         $query += "}}}}}" # close query
 
         # Setup parameters
@@ -81,7 +108,11 @@
         $out = Invoke-TibberQuery @splat
 
         # Output the object
-        $out.viewer.$homeNode.currentSubscription.priceInfo.current
-        $out.viewer.$homeNode.currentSubscription.priceInfo.range.nodes
+        @(
+            $out.viewer.$homeNode.currentSubscription.priceInfo.current
+            $out.viewer.$homeNode.currentSubscription.priceInfo.range.nodes
+            $out.viewer.$homeNode.currentSubscription.priceInfo.today
+            $out.viewer.$homeNode.currentSubscription.priceInfo.tomorrow
+        ) | ForEach-Object { if ($_) { $_ } } # return onlynodes that exists in response
     }
 }