+++ helpers/RunPester.ps1
@@ -7,7 +7,7 @@ param(
     [string] $Verbosity = 'Normal'
 )
 
-Write-Host "Running Pester: $ModuleName"
+Write-Host "Running Pester: $ModuleName" -ForegroundColor Yellow
 
 # Install/update PSScriptAnalyzer
 if (-Not $(Get-Module -Name Pester )) {
@@ -72,5 +72,6 @@ if ($errorCount -gt 0) {
     exit 1
 }
 
-Write-Host "Pester done: $ModuleName"
+# Exit script
+Write-Host "Pester done: $ModuleName" -ForegroundColor Yellow
 exit 0